requirements.txt
# Commun
pandas
xlsxwriter
pyyaml
python-dotenv
sqlalchemy
pyodbc

# Pays uniquement
pycountry
babel
geonamescache
rapidfuzz
ollama
